requirements.txt
torch
torchvision
timm
pandas
scikit-learn
pyyaml
tqdm
matplotlib
pillow
